# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Smaki CRM API Documentation",
        "version": "0.3"
    },
    "servers": [
        {
            "url": "https://api.smaki.com.ua/api"
        }
    ],
    "paths": {
...
        "/mobile/me/coordinates": {
            "put": {
                "tags": [
                    "Mobile.Me"
                ],
                "summary": "Курьерское приложение. Обновить свои координаты",
                "description": "Обновить свои координаты. Доступ для пользователей с ролью: <b>Courier</b>",
                "operationId": "b74a97f6023eb869d08630e4cd77889b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "latitude": {
                                        "description": "Широта",
                                        "type": "string"
                                    },
                                    "longitude": {
                                        "description": "Долгота",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "success": {
                                                    "type": "boolean"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "security": [
        []
    ]
}